Update marvin version with Span one and move to shared
