+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -225,7 +225,6 @@
     <Compile Include="System\Globalization\CultureInfo.cs" />
     <Compile Include="System\Helpers.cs" />
     <Compile Include="System\InsufficientMemoryException.cs" />
-    <Compile Include="System\Marvin.cs" />
     <Compile Include="System\Math.CoreRT.cs" />
     <Compile Include="System\MathF.CoreRT.cs" />
     <Compile Include="System\MissingFieldException.cs" />
